+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+
+- Added `Optional<>.hasValue()`, `Optional<>.isEmpty()`, `Optional<>.value()`
+
+### Added
 - Added `startsWith` and `endsWith` for `List`.
 
 ## [0.24] 2021-05-05